safari-600.7-branch/LayoutTests/fast/canvas/canvas-empty-image-pattern-expected.txt

CONSOLE MESSAGE: line 9: InvalidStateError: DOM Exception 11: An attempt was made to use an object that is not, or is no longer, usable.
